chore(setup): bootstrap sinatra app
A new Sinatra app has been created which lives in the `lib` directory. A
Gemfile was generated, using the `bundle init` command, to hold the
dependencies. RSpec files have been generated, using the `rspec --init`
command, in order to facilitate testing.

Shotgun has been added to allow the automatic reload of changes to the
codebase, however, it requires `config.ru` to load the correct files.