feat(orgs): display notice or repos dependent on action
If no organization is selected, a notice is displayed informing the user
that they need to select an organization in order to view its
repositories.

On selection, if that organization has repositories that belong to it,
then they are rendered as a list group.

Using a headless JavaScript driver speeds up tests, therefore
Poltergeist has been added in order to interact with PhantomJS.
Poltergeist has been setup as Capybara's JavaScript driver instead of
Selenium.