feat(orgs): allow multiple users to log in
The @@tokens class variable has been modified to be a hash. It uses the
uuid of the user as the key and looks up the token based on this. To
support this feature, the app must be of type POST - the GET route for
the app has been modified to accept both GET and POST. This ensures the
app doesn't break if loaded in a browser, it just states that the app
requires Livestax.

On a successful authentication callback - instead of calling
`window.location.reload` the app now POSTs to itself with the code
attached. This allows authentication but prevents the need to support
GET and POST.