chore(pulls): bootstrap app
This commit adds the Pull Request App, it's root route and associated
tests. An environment variable has been created for the app.

This commit also fixes naming of `repos` as `repo` in many areas. The
`validate_signed_request` method has been updated to accept the app name
in order to decode the the signed request using the correct key.