feat(logout): log out of related apps if user logs out of one
Previously, if you logged out of one app, the other apps still displayed
their when, when in fact you were no longer authenticated. We want all
related apps to be logged out when logged out of one.

This commit also adds a logout menu item to the Pull Requests App.