feat(repos): select organization based on user details
If the user is logged into Livestax, we expect a UUID, which can be used
to fetch user information, for example, the organizations that the user
belongs to.

This information is then used to select a GitHub Organization based on a
Fuzzy Search. The fuzzy search is achieved by way of of the
Damerau-Levenshtein algorithm.

Webmock has been added to stub the connection to the Livestax User API.
One can set the data of the Livestax User API by changing the User ID of
the Signed Request to the format of `key=data`, for example:

```
org=The Force&first_name=Luke&last_name=Skywalker&full_name=Luke Skywalker
```

Individual elements of the signed request can be amended by passing them
in as options.