fix(repos): return no org if user's orgs are empty
Previously, if the User did not belong to the organization the app
belonged to, the app would break as it tried to access a `name`
attribute on a `nil` organization. This fix returns an empty string if
the user's `primary_organization` is `nil`.

The `build_user_options_hash` now allows for `nil` orgs to be passed in.